[infoview] Move into its own subdir, prepare for more views.
We refactor esbuild while we are at it.
