+++ editor/code/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "files": [],
-  "references": [{ "path": "src" }, { "path": "view" }]
+  "references": [{ "path": "src" }, { "path": "views/info" }]
 }
