+++ editor/code/views/info/media/styles.css
